feat: separate create mock data in users service
